+++ crates/goose-cli/src/lib.rs
@@ -3,7 +3,6 @@ use once_cell::sync::Lazy;
 pub mod cli;
 pub mod commands;
 pub mod logging;
-pub mod recipe;
 pub mod recipes;
 pub mod session;
 // Re-export commonly used types
