+++ crates/goose-cli/src/recipes/mod.rs
@@ -1,2 +1,3 @@
 pub mod github_recipe;
+pub mod recipe;
 pub mod search_recipe;
